Files from Twenex.org.
Retrieved by Larry Masinter, August 18, 2006.  Masinter notes: "The
TWENEX files seem to have come from Sumex, or at least, there are a
lot of sumex-specific files there. The file dates are 1984, the latest
seems to be 7-Dec-84. I used 'type a' FTP for the source files and
'type I' retrieval for the binary files, since the PDP-10 is a
9-bit-byte / 36-bit-word machine."

Contents of <lisp> directory: lisp.zip

Contents of <lispusers> directory: lispusers.zip